Significantly simplified compiler; while loops
The first version was rather rudely extracted from the Herbie compiler
tools. This version does not do the value numbering that Herbie
does (because for floats only, the C compiler can figure it out).

I also added support for while loops, making it possible to translate
the C examples into the FPBench format.