core2go: initial implementation of golang compiler
This is an initial implementation of an FPCore to golang compiler based
on the core2c implementation.  It produces a package with the name of
the first argument and converts types, operators and functions to their
golang equivalents.

usage:

racket core2go.rkt package_name <input.fpcore >output.go

Known issues:
- Golang's math library only accepts float64 arguments, so usage of
  float32 inputs will cause an error on attempted compilation.